Update
Keeps track of 'father' so that if they do not preform in the arena, they may still have the ability to keep the family pool alive.

Also added more features to the main menue. Also made the list of players stay in the main menu instead of going to a different menu